exceptional handling added

+++ src/main/java/com/veersa/eventApp/service/EventServiceImpl.java
@@ -1,5 +1,11 @@
 package com.veersa.eventApp.service;
 
+import com.veersa.eventApp.DTO.EventCreateRequest;
+import com.veersa.eventApp.DTO.EventResponse;
+import com.veersa.eventApp.DTO.EventSearchRequest;
+import com.veersa.eventApp.DTO.EventUpdateRequest;
+import com.veersa.eventApp.exception.EventNotFoundException;
+import com.veersa.eventApp.mapper.EventMapper;
 import com.veersa.eventApp.model.Event;
 import com.veersa.eventApp.model.User;
 import com.veersa.eventApp.respository.EventRepository;
@@ -18,46 +24,106 @@ public class EventServiceImpl implements EventService{
 
     private final EventRepository eventRepository;
     private final UserRepository userRepository;
+    private final EventMapper eventMapper;
 
     @Override
-    public List<Event> getAllEvents() {
-        return eventRepository.findAll();
+    public List<EventResponse> getAllEvents() {
+        return eventMapper.mapToEventResponse(eventRepository.findAll());
     }
 
     @Override
-    public Event getEventById(Long id) {
-        return eventRepository.findById(id)
+    public EventResponse getEventById(Long id) {
+        Event event = eventRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Event not found with id: " + id));
+
+        return eventMapper.mapToEventResponse(event);
     }
 
     @Override
-    public Event createEvent(Event event) {
+    public EventResponse createEvent(EventCreateRequest event) {
 
         // get the user who created the event
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         User organizer = userRepository.findByEmail(email)
                 .orElseThrow(() -> new RuntimeException("User not found with email: " + email));
         // set the organizer of the event
-        event.setOrganizer(organizer);
-        return eventRepository.save(event);
+        Event newEvent = eventMapper.toEvent(event);
+        newEvent.setOrganizer(organizer);
+        eventRepository.save(newEvent);
+        return eventMapper.mapToEventResponse(newEvent);
     }
 
     @Override
-    public Event updateEvent(Long id, Event event) {
+    public EventResponse updateEvent(Long id, EventUpdateRequest request) throws RuntimeException {
+        Event existingEvent = eventRepository.findById(id).orElseThrow(
+                () -> new EventNotFoundException("Event not found with id: " + id)
+        ) ; // throws if not found
 
-        Event existingEvent = getEventById(id);
-        existingEvent.setName(event.getName());
-        existingEvent.setDescription(event.getDescription());
-        existingEvent.setDateTime(event.getDateTime());
-        existingEvent.setAvailableSeats(event.getAvailableSeats());
-        return eventRepository.save(existingEvent);
+        // Check if the authenticated user is the organizer of the event
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        User authenticatedUser = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("User not found with email: " + email));
+        if (!existingEvent.getOrganizer().getId().equals(authenticatedUser.getId())) {
+            throw new RuntimeException("You are not authorized to update this event");
+        }
+
+        if (request.getName() != null) {
+            existingEvent.setName(request.getName());
+        }
+        if (request.getDescription() != null) {
+            existingEvent.setDescription(request.getDescription());
+        }
+        if (request.getAvailableSeats() != null) {
+            existingEvent.setAvailableSeats(request.getAvailableSeats());
+        }
+        if (request.getLocation() != null) {
+            existingEvent.setLocation(request.getLocation());
+        }
+        if (request.getPricePerTicket() != null) {
+            existingEvent.setPricePerTicket(request.getPricePerTicket());
+        }
+        if (request.getIsOnline() != null) {
+            existingEvent.setOnline(request.getIsOnline());
+        }
+        if (request.getEventImageUrl() != null) {
+            existingEvent.setEventImageUrl(request.getEventImageUrl());
+        }
+        eventRepository.save(existingEvent);
+
+        return eventMapper.mapToEventResponse(existingEvent);
     }
+
 
     @Override
     public void deleteEvent(Long id) {
 
-        Event event = getEventById(id);
+        Event event = eventRepository.findById(id).orElseThrow(
+                () -> new EventNotFoundException("Event not found with id: " + id)
+        );
+
+        // Check if the authenticated user is the organizer of the event
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        User authenticatedUser = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("User not found with email: " + email));
+        if (!event.getOrganizer().getId().equals(authenticatedUser.getId())) {
+            throw new RuntimeException("You are not authorized to delete this event");
+
+        }
         eventRepository.delete(event);
 
+    }
+
+
+    @Override
+    public List<EventResponse> filterAndSearchEvents(EventSearchRequest request) {
+
+        List<Event>events = eventRepository.searchEvents(
+                request.getKeyword(),
+                request.getDateAfter(),
+                request.getMinAvailableSeats(),
+                request.getOrganizerId()
+        );
+
+        return eventMapper.mapToEventResponse(events);
     }
 }
